Get proper socket error messages on Windows
Sometimes on Windows (with winsock) errno isn't set and instead the
socket error is meant to be obtained via WSAGetLastError. At the moment
we can occasionally return errors on Windows as "No error". Use a 'get
the underlying socket error' method that falls back to errno if a socket
error isn't available.