Safeguard Type initialisations
In AllListEncoding.readValue for ArrayType and ListType if the count specified
is very large then it is most likely to trigger an out of memory error.  As
this can come from an external data source, during the SASL init, there is a
potential for a denial of service. Even though an application can completely
recover from this, the out of memory error causes a JVM to take dumps which
slows the application. The fix is to throw if the count value is larger than
the amount of data available.